commands: move obslog code to obslog.rs
